chore: split cells
